+++ api/handlers/get.go
@@ -4,7 +4,7 @@ import (
 	"net/http"
 
 	"github.com/jmpsec/osctrl/pkg/settings"
-	"github.com/jmpsec/osctrl/utils"
+	"github.com/jmpsec/osctrl/pkg/utils"
 )
 
 // HealthHandler - Handle health requests
