+++ pkg/logging/graylog.go
@@ -8,7 +8,7 @@ import (
 
 	"github.com/jmpsec/osctrl/pkg/settings"
 	"github.com/jmpsec/osctrl/pkg/types"
-	"github.com/jmpsec/osctrl/utils"
+	"github.com/jmpsec/osctrl/pkg/utils"
 	"github.com/rs/zerolog/log"
 	"github.com/spf13/viper"
 )
